Determine package date with backwards compatibility.
utils::packageDescription() was added in R 3.5.0.

https://cran.r-project.org/doc/manuals/r-release/NEWS.html

utils::packageDescription("git2r")$`Date/Publication` only works
if the package is downloaded from CRAN (which automatically adds
this field to DESCRIPTION of all submitted packages).

Thus if the package is downloaded from GitHub or other non-CRAN
source, the default is to fall back to the current year, which is
likely correct if the development version was installed.